pyproject.toml
[project]
name = "nano-slides-marp-cc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.52.0",
    "pillow>=12.0.0",
]
